requirements.txt
PyYAML==3.13
boto3